requirements.txt
google-adk>=1.14.0
python-dotenv>=1.0.0
google-cloud-logging>=3.10.0
